genportal now collects information about the duke-data-service-portal last_commit and branch and uses it to decide if it needs to generate a new portal

+++ docker/builds/genportal/genportal.sh
@@ -1,13 +1,21 @@
 #!/bin/bash
 
-cd /var/www
-git clone https://github.com/Duke-Translational-Bioinformatics/duke-data-service-portal.git
-cd /var/www/duke-data-service-portal
-npm install react-cookie
-npm install
-gulp build --type production
-rm -rf /var/www/app/portal/*
-mv dist/index.html /var/www/app/portal/index.erb
-rsync -avz dist/ /var/www/app/portal/
-echo "Generated"
+last_commit=`awk '{print $1}' /var/www/app/portal.version`
+next_commit=`git ls-remote https://github.com/Duke-Translational-Bioinformatics/duke-data-service-portal.git refs/heads/develop | awk '{print $1}'`
+if [ ${last_commit} != ${next_commit} ]
+then
+  cd /var/www
+  git clone https://github.com/Duke-Translational-Bioinformatics/duke-data-service-portal.git
+  cd /var/www/duke-data-service-portal
+  npm install react-cookie
+  npm install
+  gulp build --type production
+  rm -rf /var/www/app/portal/*
+  mv dist/index.html /var/www/app/portal/index.erb
+  rsync -avz dist/ /var/www/app/portal/ /var/www/app/portal.version
+  git ls-remote https://github.com/Duke-Translational-Bioinformatics/duke-data-service-portal.git refs/heads/develop > /var/www/app/portal.version
+  echo "Generated ${next_commit}"
+else
+  echo "nothing new"
+fi
 exit
